Fixed delete endpoints

+++ src/main/java/C043/GameVault/services/BackLogService.java
@@ -33,7 +33,8 @@ public class BackLogService {
     }
 
     public void deleteBackLog(User user, int id) {
-        BackLog found = this.getBackLogById(id);
+        BackLog found = this.backLogRepository.findByGameId(id);
+        if (found == null) throw new NotFoundException("Game not found");
         if (found.getUser().getId() != user.getId()) throw new UnauthorizedException("You don't have permissions to delete this game");
         this.backLogRepository.delete(found);
     }
